Add files via upload
I designed a radar system that can be controlled using a mobile device. This innovative system includes a display that provides real-time distance measurements of objects located in front of the user. Additionally, it features an automated sound alert system, utilizing a buzzer, which helps users gauge the proximity of the detected objects more accurately.